src/requirements.txt
git+https://github.com/sapienzanlp/extend@main
spacy
spacy-transformers
classy-core=0.2.1
crosslingual-coreference